refactor : request 패키지에 patient, visit 패키지 추가
- request 패키지 내부에서 patient, visit 패키지를 생성함으로써 각각의 request 데이터가 어디에서 사용하는지에 대한 가독성을 높임

issue : #17
